[SPARK-36303]: fix SparkCatalogException in SparkException
